requirements.txt
matplotlib==3.10.1
scipy==1.15.2
pillow==11.2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
tqdm==4.67.1
ipykernel==6.29.5
ipython==9.2.0
ffmpeg-python==0.2.0
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu128
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu128
